Added new tests for check_link
Solved issue of "No matching links" being appended to the end of the
website root. This was an mismatch between "no matching links" in
priority_match() and "No  matching links" in extract_link().

Added tests to detect the error in future and solved with an if
statement in check_link().