Added new functions
* Added tdm functions
* Added get_url function
* Added parse_html function
* Added priority_match function
* Added additional error handling to get_url.R
* Added error handling to priority_match.R
* Added test_priority_match.R
* Added extract_link and associated tests

Added new tests for check_link

Solved issue of "No matching links" being appended to the end of the
website root. This was an mismatch between "no matching links" in
priority_match() and "No  matching links" in extract_link().

Added tests to detect the error in future and solved with an if
statement in check_link().

Changed output of tdm_matrix to be corpus

Plotting wordclouds is useful, and this is much easier achieved from
a corpus rather than matrix/dataframe. An additional function would
be required to then convert corpus to dtm/tdm.